requirements.txt
django==3.1.3
django-allauth
django-crispy-forms
pypdf2
django-autoslug